Strip the "_v2" suffix from API test filenames
Rename all API test modules by stripping off the "_v2" suffix from their
filenames. For example, this:

    tests/foreman/api/test_filter_v2.py

Is now this:

    tests/foreman/api/test_filter.py

Clean up the API test stubs. Give classes names in the form `FooTestCaseStub`,
strip off redundant decorators, add docstrings and ignore some irrelevant pylint
errors.

Update the API documentation. Make `docs/api/tests.foreman.api.rst` accurate.

Fix #1493.
