+++ tests/foreman/api/test_contentviewversion.py
@@ -2,7 +2,7 @@
 from requests.exceptions import HTTPError
 from robottelo import entities
 from unittest import TestCase
-# (too many public methods) pylint: disable=R0904
+# (too-many-public-methods) pylint:disable=R0904
 
 
 class CVVersionTestCase(TestCase):
