+++ tests/foreman/api/test_foremantask.py
@@ -3,7 +3,7 @@ from requests.exceptions import HTTPError
 from robottelo.common.decorators import run_only_on, skip_if_bug_open
 from robottelo import entities
 from unittest import TestCase
-# (too many public methods) pylint: disable=R0904
+# (too-many-public-methods) pylint:disable=R0904
 
 
 @run_only_on('sat')
