+++ tests/foreman/api/test_hostgroup.py
@@ -1,32 +1,26 @@
-# -*- encoding: utf-8 -*-
-# vim: ts=4 sw=4 expandtab ai
-
-import sys
-
-if sys.hexversion >= 0x2070000:
-    import unittest
-else:
-    import unittest2 as unittest
-
-from ddt import ddt
+"""Tests for the ``hostgroups`` paths."""
 from robottelo.common.constants import NOT_IMPLEMENTED
-from robottelo.common.decorators import data
-from robottelo.test import APITestCase
+import unittest
+# (too-many-public-methods) pylint:disable=R0904
 
 
-@ddt
-class TestHostgroup(APITestCase):
-    """Testing /api/organization entrypoint"""
+class HostGroupTestCaseStub(unittest.TestCase):
+    """Incomplete tests for host groups.
 
-    @unittest.skip(NOT_IMPLEMENTED)
-    @data("""DATADRIVENGOESHERE
-        hostgroup name is alpha
-        hostgroup name is numeric
-        hostgroup name is alpha_numeric
-        hostgroup name is utf-8
-        hostgroup name is latin1
-        hostgroup name is html
-    """)
+    When implemented, each of these tests should probably be data-driven. A
+    decorator of this form might be used::
+
+        @data(
+            name is alpha,
+            name is alpha_numeric,
+            name is html,
+            name is latin1,
+            name is numeric,
+            name is utf-8,
+        )
+
+    """
+
     @unittest.skip(NOT_IMPLEMENTED)
     def test_remove_hostgroup_1(self, test_data):
         """
@@ -37,17 +31,6 @@ class TestHostgroup(APITestCase):
         @status: manual
         """
 
-        pass
-
-    @unittest.skip(NOT_IMPLEMENTED)
-    @data("""DATADRIVENGOESHERE
-        hostgroup name is alpha
-        hostgroup name is numeric
-        hostgroup name is alpha_numeric
-        hostgroup name is utf-8
-        hostgroup name is latin1
-        hostgroup name is html
-    """)
     @unittest.skip(NOT_IMPLEMENTED)
     def test_remove_hostgroup_2(self, test_data):
         """
@@ -58,17 +41,6 @@ class TestHostgroup(APITestCase):
         @status: manual
         """
 
-        pass
-
-    @unittest.skip(NOT_IMPLEMENTED)
-    @data("""DATADRIVENGOESHERE
-        hostgroup name is alpha
-        hostgroup name is numeric
-        hostgroup name is alpha_numeric
-        hostgroup name is utf-8
-        hostgroup name is latin1
-        hostgroup name is html
-    """)
     @unittest.skip(NOT_IMPLEMENTED)
     def test_remove_hostgroup_3(self, test_data):
         """
@@ -79,17 +51,6 @@ class TestHostgroup(APITestCase):
         @status: manual
         """
 
-        pass
-
-    @unittest.skip(NOT_IMPLEMENTED)
-    @data("""DATADRIVENGOESHERE
-        hostgroup name is alpha
-        hostgroup name is numeric
-        hostgroup name is alpha_numeric
-        hostgroup name is utf-8
-        hostgroup name is latin1
-        hostgroup name is html
-    """)
     @unittest.skip(NOT_IMPLEMENTED)
     def test_remove_hostgroup_4(self, test_data):
         """
@@ -100,17 +61,6 @@ class TestHostgroup(APITestCase):
         @status: manual
         """
 
-        pass
-
-    @unittest.skip(NOT_IMPLEMENTED)
-    @data("""DATADRIVENGOESHERE
-        hostgroup name is alpha
-        hostgroup name is numeric
-        hostgroup name is alpha_numeric
-        hostgroup name is utf-8
-        hostgroup name is latin1
-        hostgroup name is html
-    """)
     @unittest.skip(NOT_IMPLEMENTED)
     def test_add_hostgroup_1(self, test_data):
         """
@@ -121,17 +71,6 @@ class TestHostgroup(APITestCase):
         @status: manual
         """
 
-        pass
-
-    @unittest.skip(NOT_IMPLEMENTED)
-    @data("""DATADRIVENGOESHERE
-        hostgroup name is alpha
-        hostgroup name is numeric
-        hostgroup name is alpha_numeric
-        hostgroup name is utf-8
-        hostgroup name is latin1
-        hostgroup name is html
-    """)
     @unittest.skip(NOT_IMPLEMENTED)
     def test_add_hostgroup_2(self, test_data):
         """
@@ -142,17 +81,6 @@ class TestHostgroup(APITestCase):
         @status: manual
         """
 
-        pass
-
-    @unittest.skip(NOT_IMPLEMENTED)
-    @data("""DATADRIVENGOESHERE
-        hostgroup name is alpha
-        hostgroup name is numeric
-        hostgroup name is alpha_numeric
-        hostgroup name is utf-8
-        hostgroup name is latin1
-        hostgroup name is html
-    """)
     @unittest.skip(NOT_IMPLEMENTED)
     def test_add_hostgroup_3(self, test_data):
         """
@@ -163,17 +91,6 @@ class TestHostgroup(APITestCase):
         @status: manual
         """
 
-        pass
-
-    @unittest.skip(NOT_IMPLEMENTED)
-    @data("""DATADRIVENGOESHERE
-        hostgroup name is alpha
-        hostgroup name is numeric
-        hostgroup name is alpha_numeric
-        hostgroup name is utf-8
-        hostgroup name is latin1
-        hostgroup name is html
-    """)
     @unittest.skip(NOT_IMPLEMENTED)
     def test_add_hostgroup_4(self, test_data):
         """
@@ -183,5 +100,3 @@ class TestHostgroup(APITestCase):
         @assert: hostgroup is added to organization
         @status: manual
         """
-
-        pass
