--- tests/foreman/api/test_medium.py
@@ -1,180 +1,0 @@
-# -*- encoding: utf-8 -*-
-# vim: ts=4 sw=4 expandtab ai
-
-import sys
-
-if sys.hexversion >= 0x2070000:
-    import unittest
-else:
-    import unittest2 as unittest
-
-from ddt import ddt
-from robottelo.common.constants import NOT_IMPLEMENTED
-from robottelo.common.decorators import data, run_only_on
-from robottelo.test import APITestCase
-
-
-@run_only_on('sat')
-@ddt
-class TestMedium(APITestCase):
-    """Testing /api/organization entrypoint"""
-
-    @unittest.skip(NOT_IMPLEMENTED)
-    @data("""DATADRIVENGOESHERE
-        medium name is alpha
-        medium name is numeric
-        medium name is alpha_numeric
-        medium name is utf-8
-        medium name is latin1
-        medium name is html
-        """)
-    @unittest.skip(NOT_IMPLEMENTED)
-    def test_remove_medium_1(self, test_data):
-        """
-        @feature: Organizations
-        @test: Remove medium by using organization name and medium name
-        @assert: medium is added then removed
-        @status: manual
-        """
-
-        pass
-
-    @unittest.skip(NOT_IMPLEMENTED)
-    @data("""DATADRIVENGOESHERE
-        medium name is alpha
-        medium name is numeric
-        medium name is alpha_numeric
-        medium name is utf-8
-        medium name is latin1
-        medium name is html
-        """)
-    @unittest.skip(NOT_IMPLEMENTED)
-    def test_remove_medium_2(self, test_data):
-        """
-        @feature: Organizations
-        @test: Remove medium by using organization ID and medium name
-        @assert: medium is added then removed
-        @status: manual
-        """
-
-        pass
-
-    @unittest.skip(NOT_IMPLEMENTED)
-    @data("""DATADRIVENGOESHERE
-        medium name is alpha
-        medium name is numeric
-        medium name is alpha_numeric
-        medium name is utf-8
-        medium name is latin1
-        medium name is html
-        """)
-    @unittest.skip(NOT_IMPLEMENTED)
-    def test_remove_medium_3(self, test_data):
-        """
-        @feature: Organizations
-        @test: Remove medium by using organization name and medium ID
-        @assert: medium is added then removed
-        @status: manual
-        """
-
-        pass
-
-    @unittest.skip(NOT_IMPLEMENTED)
-    @data("""DATADRIVENGOESHERE
-        medium name is alpha
-        medium name is numeric
-        medium name is alpha_numeric
-        medium name is utf-8
-        medium name is latin1
-        medium name is html
-        """)
-    @unittest.skip(NOT_IMPLEMENTED)
-    def test_remove_medium_4(self, test_data):
-        """
-        @feature: Organizations
-        @test: Remove medium by using organization ID and medium ID
-        @assert: medium is added then removed
-        @status: manual
-        """
-
-        pass
-
-    @unittest.skip(NOT_IMPLEMENTED)
-    @data("""DATADRIVENGOESHERE
-        medium name is alpha
-        medium name is numeric
-        medium name is alpha_numeric
-        medium name is utf-8
-        medium name is latin1
-        medium name is html
-    """)
-    @unittest.skip(NOT_IMPLEMENTED)
-    def test_add_medium_1(self, test_data):
-        """
-        @feature: Organizations
-        @test: Add medium by using the organization name and medium name
-        @assert: medium is added
-        @status: manual
-        """
-
-        pass
-
-    @unittest.skip(NOT_IMPLEMENTED)
-    @data("""DATADRIVENGOESHERE
-        medium name is alpha
-        medium name is numeric
-        medium name is alpha_numeric
-        medium name is utf-8
-        medium name is latin1
-        medium name is html
-    """)
-    @unittest.skip(NOT_IMPLEMENTED)
-    def test_add_medium_2(self, test_data):
-        """
-        @feature: Organizations
-        @test: Add medium by using the organization ID and medium name
-        @assert: medium is added
-        @status: manual
-        """
-
-        pass
-
-    @unittest.skip(NOT_IMPLEMENTED)
-    @data("""DATADRIVENGOESHERE
-        medium name is alpha
-        medium name is numeric
-        medium name is alpha_numeric
-        medium name is utf-8
-        medium name is latin1
-        medium name is html
-    """)
-    @unittest.skip(NOT_IMPLEMENTED)
-    def test_add_medium_3(self, test_data):
-        """
-        @feature: Organizations
-        @test: Add medium by using the organization name and medium ID
-        @assert: medium is added
-        @status: manual
-        """
-
-        pass
-
-    @unittest.skip(NOT_IMPLEMENTED)
-    @data("""DATADRIVENGOESHERE
-        medium name is alpha
-        medium name is numeric
-        medium name is alpha_numeric
-        medium name is utf-8
-        medium name is latin1
-        medium name is html
-    """)
-    @unittest.skip(NOT_IMPLEMENTED)
-    def test_add_medium_4(self, test_data):
-        """
-        @feature: Organizations
-        @test: Add medium by using the organization ID and medium ID
-        @assert: medium is added
-        @status: manual
-        """
-
-        pass
