+++ tests/foreman/api/test_subscription.py
@@ -7,6 +7,7 @@ https://<sat6.com>/apidoc/v2/subscriptions.html
 from robottelo import entities
 from unittest import TestCase
 from robottelo.common import manifests
+# (too-many-public-methods) pylint:disable=R0904
 
 
 class SubscriptionsTestCase(TestCase):
